⚡️(frontend) specific timeout set while manifest is not ready
When users are waiting for the live to start, there is two different
cases. The first one, urls of the manifest exist and in that case, we
only fetch this url, having a timeout of two seconds is fine. The
second case is when the video doesn't have a manifest ready to read,
we then need to fetch the API to update the video store. We want to
wait another 30 seconds in this case not to load our server.