✨(back) whitelist allowed media uploaded for a shared live media
We want to control the type of media a user can use for a shared live
media. A new setting containing allowed MIME Types is created, we first
check if the given MIME Type in the initiate upload request is allowed
and we guess it if not present. When retrieved we use this exact same
MIME Type in the condition of the S3 presigned url. Doing this, AWS
will also check the MIME Type once the file is uploaded on S3 and can
reject it if the uploaded file does not have this MIME Type.