♻️(back) use permissions to compute shared_live_media serializer context
In SharedLiveMediaViewset we enrich the serializer context to know in
the context if the current request is an admin one and if it's the case
we add additional information in the serialized shared_live_media object
like the media url. This part can be refactor using the logic from the
existing permissions.