🩹(back) add consumer_site in JWT token
The consumer site is sent during the LTI connection. Models in Marsha
due to LTI’s protocol and ressources’s portability are sometimes
complex. For instance, a playlist for a given context_id can have
multiple consumer_site. It makes it impossible to retrieve a
consumer site from a JWT token. We've decided to simplify access to
this information by adding it directly in the JWT token.