♻️(back) only use is_admin in serializer context
Several ways were used to determine if the current user is an admin or
not in the serializers. To do that we injected `roles` and
`can_return_live_info` and look in the serializer for thes  values. We
must remove these keys in the serializer to keep only `is_admin` to
be less confusing and less error prone. `can_return_live_info` was
introduce in a first time to allow to return live_info for student when
they want to join a discussion. But finally we decided to inject these
informations from the front directly instead of the back application to
not disclose sensitive informations. We forget to change its name.