👔(model) liveregistrations with public or lti token
Email was previously data to decide if a user was or wasn't
registered. As we will expose the liveregistration model to log
attendances in a future commit, we've realized that the field email
is not the right data to identify a user. To simplify the process,
liveregistrations are now identifying two different scenarios, public
or LTI connection. From LTI, a user can be identified with his
`lti_user_id`, `consumer_site`, and `context_id`. At this point of
the code, a public registration can't be fetched, but two registrations
with the same email in the public mode can't be saved.