✨(back) allow to configure CORS headers
The recently added `/api/pairing-challenge` url will be used in a
browser to a not recognized domain. This requests are rejected because
CORS headers are not activated in Marsha. We added the
django-cors-headers package responsible to manage them. The
configuration is not fully made to not block the strategy you want to
use. You have to decide on your marsha instance the strategy you want by
configuring one of the following setting `CORS_ALLOWED_ORIGINS`,
`CORS_ALLOWED_ORIGIN_REGEXES` and `CORS_ALLOW_ALL_ORIGINS`.
The full documentation is available here
https://github.com/adamchainz/django-cors-headers#configuration