✨(backend) display_username and anonymous_id in LiveRegistration
Users from LTI connections don't necessarily have a username.
Users watching the video from public access are anonymous. We
want to make sure, we have a unique display name by video
regardless LTI or public connection, so users in the chat or in
the attendance list are whole identifiable. The field
anonymous_id is introduced to be used for public connection to
be able to identify users.