🐛(back) load development urls only when DEBUG is true
In the DevelopmentViewset we use packages required only in development
environment. But this module is imported in marsha.urls and we have a
ModuleNotFoundError in an environment have DEBUG set to False. To avoid
this error, an `urls` module is created in marsha.development and is
loaded in marsha.urls using django.urls.include only when DEBUG is True.