✨(frontend) anonymous_id is added to register to a scheduled video
For public connection, we had no way of remembering the user as the
token contains no data. We now will send a uuid in the request to
register a user to a scheduled video. This uuid will be saved and
be used later on to identify the user. At this point, the uuid is
randomly generated, a future commit will create a uuid for each user
accessing the front-end for the first time.