✨(backend) liveregistration with public connection
Now we have added an anonymous_id key to deal with the public
connection, we can recognize users from a public connection. So
far any user from a public connection couldn't be recognized
and no treatment was done in this specific case, only LTI users
were identified. We adapt the code so users from public video
can be identified.