✨(frontend) add chat's changes to integrate new authentification feature
Add a new feature to the chat. An incoming user is now automatically login as
an anonymous one when he arrived on the chat. He can display messages but if he
wants to join it and be able to write his own he has to authenticate himself.
A dedicated overlay component is added to fulfill this role.