✨(back) add middleware checking JWT for a websocket connection
To identify a user when he tries to connect to the websocket, we want to
use the already existing JWT token. It is not possible to use the
existing mechanisms used with DRF and it's not possible to add
additional headers when a connection to a websocket is made. The token
must be added to the `jwt` query string argument and if valid the token
is added to the consumer scope. Unfortunately it is not possible to
check at this point the resource_id in the token is the same as the one
used in the route path.