✨(back) add a video consumer to marsha.websocket
The first consumer we need is one related to the Video model object. The
consumer is responsible to accept or not the websocket connection. In
this consumer we check if the resource_id in the token and the id in the
route path are the same and if the video is really exists in the
database. For now this consumer does not manage received message from
the websocket. It only listen the channel layer and send all clients the
serialized video if it receives a `video_updated` message type in the
channel layer.