✨(back) override channel_redis channel layer
The channel layer provided by channel_redis use msgpack to serialize and
deserialize messages. We have a problem when a video is serialized. It
is impossible then to deserialize it because in the urls the resolution
can be used as key and those keys are integer. Msgpack is unable to
deserialize this kind of key. As suggested in the issue[1] open on
channel_redis git repository we implement our own serialize and
deserialize methods.
1: https://github.com/django/channels_redis/issues/287