✨(Mails) convert mjml files to html and text
Mjml is a responsive email framework designed to reduce the pain of coding
A responsive email. Editor exists online and emails can be built easily by
anyone. Mjml can’t be used directly to send emails, it needs to be converted
into text and html. We create a new container that is in charge of this
service and convert mjml files to html and text.

Resolve https://github.com/openfun/marsha/issues/1244