🎨(make) change help task to better display self documenting makefile
The help task was self documenting the makefile as expected. It should
display first the name task and then its documentation suffixed by ## in
the makefile. This is the same help task used in other openfun projects
but the usage of MAKEFILE_LIST is modified because we include other
files in our makefile and this variable will contains all file name
included. So we have to use only the first word and not all.