✨(front) add grouped messages and join & left label in the chat
The way the chat works changes. Consecutive messages (same sender in a short
period)  are displayed together and a label is displayed when a user joins or
leaves the chat. Also, the spinner displayed while history message are received
is now waiting for all history message to have been received, before going
hidden.

CCD