✨(back) create management command to send reminders
When webinars are scheduled, users can register. They might have
registered a long time ago. For them not to forget about it, we
send them reminders when approaching the beginning of the webinar.
Webinars have a date planned, but instructors can start the session
earlier. We want to inform users that live has started before.
Emails templates at this point are in a very basic version.

Resolve https://github.com/openfun/marsha/issues/1261