🐛(back) use Video resource name for the type sent in the Websocket
Everywhere in the application we use the RESOURCE_NAME constant on each
model to identify the resource type. This is the value sent to the front
application to determine the resource we are working on. In the websocket
message we must change to this constant to avoid error like using the actual
singular form which is not recognize by the front application.