📱(frontend) add responsive supports to student live layout
Desktop layout can not be used on small screens like smartphone. Update live
layout, on small screens actions are group by at the bottom of the screen,
live panel when open takes the whole screen and player or jitsi is hide behind
to keep the live playing.
When the smartphone is rotated, player takes the whole space and no actions are
accessible.