♻️(front) use websocket instead of polling the API during live workflow
To determine the live state in the front application we used to fetch
the API and detect if the live state has changed. In many part of the
application this polling system was used. By using the websocket we can
remove all of them. But we still need to deal with the HLS manifest
state. We still need to test if the manifest exists and once the live
resume we still need to check if the manifest has new frame to display.
All this check have been moved to the video player directly