♻️(front) use @converse/headless instead of converse.js
We are not using converse UI anymore and installing the whole
converse.js package is heavier. We only want to use @converse/headless
package. In order to use, we have to copy an entrypoint front the
converse.js project that loads the library and add it to
window.converse
To ease its usage and modification we keep it in javascript instead of
typescript.