✨(front) hide StudentLiveLayout actionElements while live not started
While the live is not really started (the manifest does not exists) the
action elements must be hidden. To detect this special state we use the
useLiveStateStarted store. A special case is managed, when the student
is on stage, we don't check anymore the store state. To be here it must
use the action elements itself.