🔨(command) add a development command to simulate emails
Changing emails templates and checking the results without any
tool is complicated. To simplify this task, we create a command
that can only be used in the development environment and that
sends samples of the four distinct reminders mails. This way
we can easily check the final layout of each mail.