♻️(front) use websocket instead of polling timed text to update state
When a timed text track upload_state is updated, a message is dispatched
in the websocket and the zustand store is updated. The UI will be
updated accordingly, there is no need anymore to poll the timed text
track endpoint to detect when the state is updated. To simulated this
update, the jest websocket mock is used to send an updated timed text
track.