♻️(front) use websocket instead of polling thumbnail to update its state
When a thumbnail upload_state is updated, a message is dispatched
in the websocket and the zustand store is updated. The UI will be
updated accordingly, there is no need anymore to poll the thumbnail
endpoint to detect when the state is updated. To simulate this
update, the jest websocket mock is used to send an updated thumbnail.