🔥(front) fix stop processing unknown stanzas
When receiving stanzas, it can happens some of them are not Messages, IQ or
Presence (smash websockets usings stream stanzas by example). These stanzas are
unrecognized by the front and an error is raised. This is because in the actual
implementation, when a stanza is unknown, chatPlugin.ts raises an error. This
commit changes this part, and makes the plugins stop processing unknown
stanzas (because, at the end of the day, purpose is to be able to process
recognized stanza and no more).