♻️(frontend) no longer use a layer to present live pause alert
Pause component mounts a layer to display the alert for the student.
But there is an issue on mobile display since the layer may be rendered above
the panel containing the chat or the viewer list.
We rather use a stack to render the pause alert in front of the player.