🐛(frontend) display waiting screen only when live has not started yet
Currently, when a live is paused then resume, the student is presented with the
waiting screen as if the live has not started yet.
We now render the waiting screen only as long as the live has not started once
(aka until the manifest can be fetch). During a pause and when the live is
restarting after a pause, the student is advertise only with the pause screen.