✨(backend) add video api endpoint to store live participant lists
When participants have joined a live, their list is lost if browser is reloaded.
When participants ask for joining, their list is lost too.
We need a way to persist those lists.
Two video API endpoints have been added to store them, and has been added to
the video serializer.