✨(front) disable jitsi deeplinking
Jitsi deeplinking checks if there is a more adapted application than the
browser to suggest depending on the device you are using. If you are on
a mobile device it will suggest you to download the mobile app and use
it. We want to prevent this behavior and keep the user in the browser.