✨(front) address prosody's answer when entering a displayName
When a user enters a valid username (valid syntax and length) and validates it,
a spinner will be displayed in place of the svg's button. Then, if the server
takes too much time to answer or if the username is already used in the chat,
an error will be displayed and the user will have to retry.