♻️(back) configure with settings the default lti locale
In a LTI request, the locale is defined using the property
launch_presentation_locale in the request. When this one is missing,
since now we were fallbacking with `en` value. We want to configure this
default value and we created a new setting for this named
DEFAULT_LTI_LAUNCH_PRESENTATION_LOCALE