✨(back) accessing videos from mails for scheduled webinars
Mails are sent for scheduled webinars. We want to add a link
to the webinar directly in the mail for users to easily access
the video. Registration might have been done inside a LTI context
or not as video might be public. The video will be displayed
directly on marsha and a JWT token using the information set in
the liveregistration record will be generated.