🐛(back) allow to configure converse persistent store
Converse allows to use multiple persistent stores. By default the one used
is IndexedDB. This is a modern persistent store avaiable in browsers but
has some caveats. Browsers have more and more mechanisms to protect
user private life, like not allowing the usage of IndexedDB database
when the script using it is not on the same domain (cross domain). In
that specific case (private browsing or by default with Brave), converse
is not usable and does not fallback on another storage. For this, we
want to configure the persistent store used by converse and we can
configure it by changing the XMPP_CONVERSE_PERSISTENT_STORE setting.