✨(backend) simplify attendees list sent by BBB
When we fetch infos about a meeting, we format sent xml to json.
If only one attendee exists, it was sent as an object, not in a list.
Furthermore, useless keys makes looping to them harder.
Now, they are always in a list directly at the root of the info.