✨(frontend) prevent multiple join to BBB meetings
When a user joins a meeting, we send to BBB a unique user ID.
We loop through all meeting participants to check if the current
user is already connected.

As participants list contains disconnected users, we have to check
if those are currently listening, speaking, or sharing their screen.
This permits to ignore disconnected users in our verifcation.