♻️(front) factorize checkLtiToken function
In the StudentLiveRegistration component a function checkLtiToken was
created in order to check if the current JWT token is a LTI one or not.
We will need to use this function in other part of the code so we create
a dedicated module in the utils directory. Doing this we spot that the
email property was not present anymore in the JWT token, it was moved
under the user property.