🐛(frontend) show harvesting infos at the end of a live
When a live is stopped a process to harvest the video is triggered and takes
some times but the layout is not updated to provide feedback.
Fix this by using the live layout only until the harvest process has started and
provide infos when the live has stopped and the process to harvest is about to
begin.