🐛(frontend) hide registration email input for lti students
If a student uses marsha using a LMS (aka he is a lti user) and this user has
already an email, we don't want him to be able to register to a live using an
other email.
Therefore, we no longer mount the input in the registration form for this type
of user.