⚡️(docker) enable buildkit by default to build project images
Images built using the make build command are built with docker-compose.
In order to use buildkit with docker-compose we have to use two
different environment variables. The first is COMPOSE_DOCKER_CLI_BUILD=1
to instruct docker-compose to use the docker CLI to build the image and
the second is DOCKER_BUILDKIT=1 to enable the usage of buildkit