🐛(docker) do not "install" marsha python package
The `marsha` package does not need to be installed here, as we only need
dependencies installation. This also prevents the use of the "marsha
installed package" instead of the later copied full `/src/backend/`
project.
The `dev` and `e2e` images were still installing the marsha package
so we explicitly uninstall it.

This fixes the following side effect too:
Files of the marsha Python package where copied directly into the
/builder/ directory making the setup.cfg at the same level as the
Django applications' module, resulting in a wrong interpretation of the
package structure: each Django application was considered as a Python
package (thanks to the find: automatic attribute).
The issue raised on my side was the markdown application overwritting
the markdown original package... For instance, the bug was visible in
the back-builder image, were the bbb application was directly inside
the /install/lib/python3.10/site-packages/bbb/ folder and so on.