✨(frontend) auto-set anonymousId when contained in the JWT token
When users subscribe to a scheduled webinar, they will receive emails.
These ones contain direct access to the video. For public
resources, thanks to this generated link, we are able to detect the
user that is watching the webinar, reading the JWT token. In this
specific case, the anononymousId generated during the registration
is added to the JWT token. We can then read this data and auto-set
the anonymousId in the localstorage with this value.