🔧(docker) restrict what is copied in the production image
We analyzed the production image to detect if files are copied in it and
are not needed to run in production. We updated the .dockerignore file
to ignore files used in the development process. In the Dockerfile, in
the final layer we only copy the backend application. The front sources
are not needed anymore, they have been built and are copied in the
static repository.