🐛(frontend) fix converse initialisation
Converse is initialize when 'components/Chat/index.tsx' file is loaded but we
can have scripts using converse from window even if converse has not been
initialized yet.
Especially, current implementation works only because we are importing Chat file
in base routeur for the application.
To fix this issue, a new component is created to be responsible for the
initialization of converse and mount component using converse only when this is
done.