🔨(frontend) add react query devtools
As react-query is used in the project we can enable this debug tool. As
stated in the documentation, this will not be added to the production
bundle.

See https://react-query.tanstack.com/devtools#import-the-devtools
```
By default, React Query Devtools are only included in bundles when
process.env.NODE_ENV === 'development', so you don't need to worry
about excluding them during a production build.
```