💩(frontend) dirty hack to keep jitsi infos
When a student joins a live discussion, he gets the jitsi infos through
a XMPP message and can join the meeting.
But when the video is updated, those informations are lost.
This is a hack that needs to be removed when an anonymous id will allow
to send to participants in discussion all the informations they need.