♻️(back) admin can access all livesession belonging to a video
We want to let an administrator or instructor to access to all the
live_session belonging to a video and not only to the consumer_site and
context_id he is looking the video. The same way a display_name is
unique to a video, an instructor should access to all live_session
belonging to a video.