🐛(front) keep input register email display
If the user has no email in its JWT token, he can enter one by using an
input. This input was removed after he enter a first caracter because we
were checking if the entered value was not empty to display this input
or not. We have to do this check on the initial email value and not the
one entered by the user.