⚰️(front) removal of unused joinDiscussion components
JoinDiscussionAskButton and JoinDiscussionLeaveButton are unused in the
project.The JoinDiscussionAskButton asks for a username which is already the
purpose of the InputDisplayNameOverlay component in the chat. It also displays
different messages depending on if the user is asking, accepted or rejected
from the stage, which is already the role of the LeaveJoinDiscussionWrapper
component. In the same way, JoinDiscussionLeaveButton is responsible of
managing an accepted user user who wants to leave the stage, and it is also
already adressed by LeaveJoinDiscussionWrapper component.