♻️(front) refacto of the jitsi route and associated component
When a user was going on stage, he was redirected to a new route  (
PUBLIC_JITSI_ROUTE) and the PublicVideoLiveJitsi component was displayed. Now,
when a user goed on stage, the StudentLiveWrapper component is updated because
the accepted attribute in useParticipantWorkflow  store passes to true.
Depending on the value of this attribute, StudentLiveWrapper component displays
a videojs player or the jitsi one.