🐛(back) expose XMPP info for a live once this one started
When a live is created, it has a IDLE live_state and the XMPP room is
not created at this time. Returning the XMPP info in the serializer is a
mistake and lets think that the XMPP room can be used by the front
application which is not the case.