✨(front) add a vertical slider to resize panel on live view
Add a vertical slider between the video and the right panel, in order to beable
to re-size the panel to fit the desired size ratio of the user. Furthermore, a
button is added to be able to open and close the chat. Finally, in order to be
able to know when the xmpp is ready (for the LiveVideoResizer), LiveVideoLayout
has a new prop added, to be able to pass this information from its parent to
its LiveVideoResizer child.