🐛(back) continue stopping live even if input waiter fails
The medialive input waiter can fails when we try to delete the aws
stack. If the max attempt is reached, a WaiterError is raised, the
endpoint return a 500 and then it is not possible any more to stop a
live. To fix this issue, the WaiterError exception is catch and sent to
sentry and the stop endpoint continues its job.