✨(frontend) ask for display name on anonymous join discussion request
Since anonymous id introduced in the application, students are automatically
connected to XMPP and can join the discussion without having to set a proper
name.
To ask the name we reuse and factorise the join chat functionality, both join
chat and join discussion will use a single store, modal and form to ask the
user his display name.