♻️(back) resource title is no more mandatory
Resources title was mandatory. Due to this, the title was set with a non
human readable text for a LTI initial request. The display title is in
most of case long and not understable (like a UUID). We choose to set
the title not mandatory, left unchange when the resource is created with
a LTI request.