🔥(email) remove email when webinar is already started
We were sending an email if the webinar had started before
its scheduled time. We finally decided not to warn users
in this scenario anymore. Teachers might do some tests and
we don't want to send an email to users when the live is
not actually ready. All code has been deleted from the
project.

Fixes #1396