🐛(frontend) live layout height bug on small devices
When using the live feature, teachers would face an issue preventing them from
using any actions. Any css update (like overing a button) would increase the
height of the document.
This is due to the height of the live layout set to 100vh but parent component
was still mounting content under the live layout.
The fix consists on using the live layout to mount the content under the live,
therefore handling properly the height of the content and scrolling the live
layout if needed.