🐛(backend) disable is_public check for meeting
When we make a get request for a meeting, as we are not in a LTI
context, we check if the meeting is publicly available through an
is_public attribute.
But as a meeting would never be public it has no is_public attribute,
so a new NonPublicBaseLTI class has been added which blocks both
get request and the is_public check.

All LTI resources which can't be public should inherit from this class.