🎨(frontend) button hover style reworked
Live actions button style was buggy and we would not distinguish a hover button
from a selected one not hovered.
Style is fixed to have a single style for hover button with a dedicated color
and if the button is not hovered then handle the style according to its
configuration.