⚰️(back) remove video constraint video_unique_idx
The video_unique_idx constraint based on the unicity of the couple
lti_id and playlist_id is a legacy constraint from when the lti_id was
essential to the way we let users interact with video and documents
through LTI requests. Since the lti_id is optional and on Moodle it can
not be unique anymore.
This commit complete this one 9d670b4b46cce5ca9f26f6ea62ce083e4bb05103