🐛(front) patch video to publish it once harvested
Updating the video using a PUT request is not the better way to update a
video. The title is null when the video is created but we force it to
not be null when the user tries to update it. Using a PUT request, we
send the entire object and, sometimes, the title can be null. To fix
this issue, we decided to use a PATCH request and send only the needed
fields in the body.