✨(front) set language in livesession from JWT token
Views of Marsha are mostly displayed from the front. The language
of the content is set depending on the locale sent in the JWT token.
We send this value to the livesession endpoints so this language is
common between the front and the back. The back will then use this
value for instance to send email asynchronously.