✨(front) add a new widget to make live publicly available
This widget allows the instructor to make the live public or not. In case he
decides to make the live public, this widget also displays the public url of
the video, to the non-LTI users to access it. This url can be directly copied
in the instructor clipboard, if he clicks on the dedicated button.