🐛(frontend) live vod thumbnail width
Live that has been recorded and published display a default thumbnail too big
and expanding on the right side outside of the bounding box.
The main regression is probably in the generation of the thumbnail for a
recorder live but frontend should not rely on image width to be rendered
properly, so I added a max width for the thumbnail container.