♻️(back) add settings CLOUDFRONT_SIGNED_PUBLIC_KEY_ID
To use the new public key from AWS cloudfront we have to use its id with
botocore cloudfront signer class. Previously the setting
CLOUDFRONT_ACCESS_KEY_ID was used but this one is not relevant anymore.
We choose to remove it and create instead a new one called
CLOUDFRONT_SIGNED_PUBLIC_KEY_ID and should contains the public key id
generated by terraform and can be found in the terraform output.