✨(front) push video live attendance
When a webinar is watched by a student, we want to record his
assiduity to show a live report of attendance. When did the
student start to watch the live? Is the sound on and volume
audible? Does he want to pause the live?... Any information
that will help the instructor to adapt his live and give him
clues to feel more in a real class.