🩹(front) fix calendar display
The calendar was displayed under the date input and part of it was sometimes
appearing outside the iframe, making it unusable. The current behavior is now
to be always displayed above the dateinput and given our application layout, it
guarantees it will always have enough space to display.There was also a
problem with the calendar button on the date input, which wasn't disabled when
the prop 'disabled' was on. It is now fixed by disabling pointerEvents on its
parent component.