✨(back) list of live attendance by user for a video
With a common delay for all users, we record information
when a student is watching a live. This is saved depending
on its own line of time : students might arrive at different
time and during the live a student might not be present all
the time.
When the instructor requires to list the attendances, we
want to provide the results with a common line of time for
each student. For this purpose, we have to build a common
timeline depending on the length of the video and check
for each student, if we have known information.