✅(front) change select for rendered mermaid graph
The mermaid library now add a title to the rendered SVG. The title is
rendered twice. Once in the svg title and once in the SVG body as a text
attribute. We have to change the selector to test it. We choose to use
the img role and check for its title.