⚡️(back) create fallback cache strategy when redis is down
When redis is down and used as cache backend we want to fallback on an
other backend cache to avoid denial of service. For this we have copied
the fallback cache used in the openfun/richie project and we adapt it to
our use case. For example we don't need to clear any cache if redis is
up again, the cache we are using has a lifetime.