♻️(backend) define marsha admin site as default
The previous implementation prenvented to really
use the admin autodiscover for dependencies because
they register against the default Django admin site.
Adding the dependency admin modules was really painful
as we must ensure all the applications are loaded
before trying to register any admin module, and this
is not how how Django expect it to work
(except for stand-alone scripts).

This new implementation defines the Marsha admin site
as the default one, making all admin modules to be
properly autodiscovered and added to it.