♻️(back) use custom policy to generate cloudfront signed url
Since the beginning of marsha we used predefined signed url to request
cloudfront. So we needed to generate a signed url for each resources we
want to fetch. Signing a url is a low operation and when we introduced
shared_live_media resources we had a bottleneck on the video and
shared_live_media serializer. To fix this issue, we decided to use a
custom policy[1]. This custom policy target at once all the resource
belonging to a given video. We sign this this policy once and then we
catch all the needed parameters to use them with all the resources
related to this video.

[1]: https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/
private-content-signed-urls.html#private-content-choosing-canned-custom-policy