👔(backend) provide only base url for new resource on lti select
When we add a new resource through LTI select (deep linking) we want to
create it.
So, we don't need to generate a new uuid in this context.
We only send base url, and the front will concatenate it with the
resource id newly created to send it back to the LMS.