➕(backend) add dependencies for SAML auth
We add Python Social Auth dependency along with the Django adapter and its
configuration.
The `signxml` is only added for development purpose to easily sign our
SAML responses in testing or developping environments. Even if it's
production ready, we don't rely on it since Python Social Auth provides
eveything we need.