✨(backend) add Renater XML parser
This adds new module which allow to read the Renater's federation metadata
and convert it to Python Social Auth usable objects.
We highly rely on the way Python Social Auth works and reuse their own helpers.
The parsed metadata is stored in cache (will be Redis cache on production), and
a management command will fill the cache on a daily basis (needs to be added to
crontab).