✨(front) fetch video data after websocket reconnection
When the websocket reconnect to the server after a disconnection, we
fetch the video data in order to refresh them in the zustand store.
Doing this, we will not miss data between disconnection.
Also, the robustwebsocket library type are wrong. We choose to create
our own and overrde the reconnection strategy. By default, the lib does
not reconnect to the server after three attempts. We do not limit the
number of attempt anymore but we increse the delay to 500 ms after 10
attempts.