🐛(back) force to return None in Thumbnail serializer if deleted
The django-safedelete apps does not manage OneToOneField relation so a
deleted thumbnail is return in the VideoSerializer even though the
thumbnail is deleted. To simply resolve this issue we check in
ThumbnailSerializer.to_representation if the thumbnail is deleted and
force to return False if deleted.
An issue is open on the makinacorpus/django-safedelete to track this bug
https://github.com/makinacorpus/django-safedelete/issues/211