🐛(back) change thumbnail video relation to ForeignKey field
The relation between thumbnail and video was a OneToOneField and it is
exactly the behavior we want. Unfortunately the app djang-safedelete
does not manage it and it is then impossible for us to delete a
thumbnail belonging to a video. The thumbnail is always return. To fix
this issue we decided to change the relation with a ForeignKey field and
to mimic the behavior of a OneToOneField by adding a UniqueConstraint
and managing the uniqueness of the resource in the VideoSerializer.
Moreover, we must keep the thumbnail instance in the VideoSerializer to
avoid to query the database again and again to fetch the instance every
time we need it. Without this, the number of queries is growing.
An issue is open to track this bug
https://github.com/makinacorpus/django-safedelete/issues/211