🐛(back) migrate video upload_state and live_state data
In the migration 0049 we apply some modification that were pending since
a while to finish the beta process without any pending migration. Doing
this, we missed to migrate some data in the database. The video
upload_state and live_state choices have changed. We must migrate the
corresponding data to have a coherent database.