✨(front) add notifications for manage on-stage request
When an instructor received an on-stage request, if he is not on the viewers
stage, a notification pops up and notifies him. The notification also allows
him to directly access to the viewers tab. Finally, when live panel is opened
but not on the viewers tab, a bell with the number of on-stage request is
displayed on the viewers tab's title.