✨(front) fetch livesessions using react-query
We remove the usage of simple fetch in a function to fetch livesessions.
Instead we use react-query in all concerned components
and work with its cache to not make extra not needed requests. This is
the case if the user is already registered or already have an existent
livesession. In that case there is no need to fetch again and again the
livesessions. If there is no livessessions, we need to create one in
that case and we must invalidate the cache to refetch the newly created
livesession.