✨(frontend) create a custom render method for testing
Create a custom render method to use in our tests wrapping the rendered
component accordingly with the application architecture.
The render method uses under the hood react testing-library render method and it
is made to be used in a similar way. The main benefit is the wrapping of the
rendered component in :
- grommet main context (grommet)
- grommet specific responsive context (grommet)
- react router (react-router-dom)
- react i18n context (react-intl)
- breadcrumb context (custom one)
- query context (react-query)
It also added to the rendered tree :
- styles from styled-reboot for styled-components
- toast and its hack to clean up toast (react-hot-toast)