🏷️(front) better type react-query fetchList
In the fetchList method we want to clean the querystring used to avoid
having some with an undefined value. To do this clean, we have to cast
the queryKey parameter and we also want to avoid this. We enhance the
types used by this function to properly work with its parameters.