🔨(front) allow to disable chrome sandboxing with puppeteer
Chrome uses multiples layers of sandboxing and the host should be
configure to work with these sandbox. Since ubuntu 22.04 I found no way
to configure chrome sandboxing and I am not able anymore to run the all
test suite without disabling the sandbox. Disabling the sandbox is
optional and you must use an environment variable for this. To disable
it, the environment variable to use is DISABLE_PUPPETEER_SANDBOX and you
must set it to 1 to disable sandboxing.