✨(frontend) create a store to expose video and live
Video and live are resources used widely in the application but are not exposed
in any context, resulting on the video being propagated through props in many
components.
Adding a react context to expose the video and the live once checks have been
made to ensure the availability of the video or the live.