🐛(aws) upgrade cairo version used in lambda convert
To convert a PDF into an SVG, we use the tool pdfToCairo made by
poppler. Cairo is responsible to the SVG part. In the lambda, an old
version of cairo is used with a known bug that "mangled" som SVG. To fix
this bug we have to build and install a newer version of Cairo fixing
this bug.